Refactor worker backend drivers into dedicated module
